+++ NEWS.md
@@ -1,2 +1,6 @@
+0.2.0
+* `tibble` fix by Danilo Zocco (PR #1)
+* Removes {dplyr} & {purrr} heavy dependencies
+
 0.1.0
 * Initial release